NOTE: Most files are changed during the refactoring of model layer. Please sync your local codes before proceeding with any change.
DID: 
1) re-generate all Model classes and Column-name constants according to original ecshop table/columns names
2) primary key field is renamed to pkId
3) added a Long type longId field and corresponding Dao methods, it's not mandate for all but some entities such as Goods need a numeric type ID field, in order be behave correctly in Javascript. see GoodsWrapper.getGoodsId()

TO:
make it easier to implement web front application reusing the original ecshop template

+++ trunk/JCommerceGae/core/com/jcommerce/core/test/TestGoods.java
@@ -22,7 +22,7 @@ public class TestGoods extends BaseDAOTestCase {
 			Goods goods = new Goods();
 			goods.setKeyName("xxx");
 			goods.setGoodsName("abc");
-			goods.setGoodsNumber(123);
+			goods.setGoodsNumber(123l);
 			
 			String id = manager.txadd(goods);
 			Long ld1 = goods.getLongId();
@@ -36,7 +36,7 @@ public class TestGoods extends BaseDAOTestCase {
 			System.out.println("ld1="+ld1+", ld2="+ld2);
 			String name = goods.getGoodsName();
 			System.out.println("name="+name);
-			int gn = goods.getGoodsNumber();
+			long gn = goods.getGoodsNumber();
 			System.out.println("gn="+gn);
 			assertTrue(ld1.equals(ld2));
